requirements.txt
google-generativeai
anthropic
openai
matplotlib